This is the initial draft of a Node.js wrapper in typescript. Approximately 200 tests have been stubbed out, based on the tests in the java library - but they simply call assert.fail()
In addition, several known issues exist around wallet types and
callbacks, particularly via return parameters.

It is my hope that we can make whatever major modifications are
required in this branch, and come to consensus on a solid solution
which supports typescript/express and javascript on node - all
from a common code base.

Once we have a viable solution that fits everyone's packaging and
structural needs, we can migrate to libindy for ticket-driven
development with feature-and-fix oriented updates, rather than the
large structural issues we still have to deal with.